+++ deno.json
@@ -11,7 +11,6 @@
     "test:e2e": "deno task -f @e2e/node test",
     "test:code": "./run-tests.sh"
   },
-  "imports": {},
   "lint": {
     "rules": {
       "exclude": [
