+++ docs/site/docs/home/100-components/100-components.md
@@ -9,6 +9,7 @@
 * [API](./103-api.md)
 * [Contracts](./105-contracts.md)
 * [Orchestrator and Processes](./106-processes.md)
+* [Environment Variables](./199-environment-variables.md)
 * [Block Explorer](./107-explorer.md)
 * [Batcher](./108-batcher/1200-overview.md)
 * [Database](./109-database.md)
