+++ e2e/client/node/.env.e2e
@@ -1,0 +1,17 @@
+# SAMPLE ENV FILE
+#
+# AUTOMATIC LOADING:
+# This file will be loaded if EFFECTSTREAM_ENV=dev is set 
+# In general .env.{EFFECTSTREAM_ENV} will be loaded for different envs
+#
+# SECURITY:
+# Do not commit your .env.xxx files
+#
+# READING ENVS: 
+#   ENV.getString("SECRET_TEST") => "test"
+#   ENV.getNumber("SECRET_TEST") => 0
+#   ENV.getBoolean("SECRET_TEST") => false 
+# NOTE: Only case-insensitive ["true", "t", "1", "yes", "y"] are considered true
+#
+# This is used for e2e testing.
+SECRET_TEST=test
