+++ e2e/client/node/.env.quickstart
@@ -1,0 +1,16 @@
+# SAMPLE ENV FILE
+#
+# AUTOMATIC LOADING:
+# This file will be loaded if EFFECTSTREAM_ENV=dev is set 
+# In general .env.{EFFECTSTREAM_ENV} will be loaded for different envs
+#
+# SECURITY:
+# Do not commit your .env.xxx files
+#
+# READING ENVS: 
+#   ENV.getString("API_SECRET_KEY") => "12345678"
+#   ENV.getNumber("API_SECRET_KEY") => 12345678
+#   ENV.getBoolean("API_SECRET_KEY") => false 
+# NOTE: Only case-insensitive ["true", "t", "1", "yes", "y"] are considered true
+#
+API_SECRET_KEY=12345678
