+++ e2e/client/node/deno.json
@@ -10,9 +10,12 @@
     // separate from node:start so we can type-check before starting other services
     "check": "deno check --unstable-raw-imports src/main.ts",
     // Open chrome://inspect to see the inspector
-    "node:start": "deno run -A --inspect --unstable-raw-imports src/main.ts",
-    "dev": "deno task -f @effectstream/tui clean && NODE_ENV=development deno run -A --check --unstable-raw-imports scripts/start.ts",
-    "test": "deno run -A --unstable-raw-imports --check scripts/e2e.test.ts"
+    
+    "node:start:quickstart": "deno run -A --inspect --unstable-raw-imports src/main.ts",
+    "node:start:e2e": "deno run -A --inspect --unstable-raw-imports src/main.ts",
+
+    "quickstart": "deno task -f @effectstream/tui clean && EFFECTSTREAM_ENV=quickstart NODE_ENV=development deno run -A --check --unstable-raw-imports scripts/start.quickstart.ts",
+    "e2e": "EFFECTSTREAM_ENV=e2e deno run -A --unstable-raw-imports --check scripts/start.e2e.ts"
   },
   "imports": {
     "pg": "npm:pg@^8.14.0",
